# [Refactor] Improve errors.Join performance