--- a/example/client/go.mod
+++ b/example/client/go.mod
@@ -11,7 +11,7 @@ replace (
 	golang.org/x/crypto => golang.org/x/crypto v0.8.0
 	golang.org/x/net => golang.org/x/net v0.9.0
 	golang.org/x/text => golang.org/x/text v0.9.0
-	google.golang.org/genproto => google.golang.org/genproto v0.0.0-20230403163135-c38d8f061ccd
+	google.golang.org/genproto => google.golang.org/genproto v0.0.0-20230410155749-daa745c078e1
 	google.golang.org/grpc => google.golang.org/grpc v1.54.0
 	google.golang.org/protobuf => google.golang.org/protobuf v1.30.0
 	gopkg.in/yaml.v2 => gopkg.in/yaml.v2 v2.4.0