--- a/internal/errors/errors_benchmark_test.go
+++ b/internal/errors/errors_benchmark_test.go
@@ -67,6 +67,7 @@ func randStr(n int) string {
 		cache >>= rs6LetterIdxBits
 		remain--
 	}
+	// skipcq: GSC-G103
 	return *(*string)(unsafe.Pointer(&b))
 }