--- a/internal/slices/slices.go
+++ b/internal/slices/slices.go
@@ -25,3 +25,22 @@ func Sort[E constraints.Ordered](x []E) {
 func SortFunc[E any](x []E, less func(left, right E) bool) {
 	slices.SortFunc(x, less)
 }
+
+func SortStableFunc[E any](x []E, less func(left, right E) bool) {
+	slices.SortStableFunc(x, less)
+}
+
+func RemoveDuplicates[E comparable](x []E, less func(left, right E) bool) []E {
+	if len(x) < 2 {
+		return x
+	}
+	SortStableFunc(x, less)
+	up := 0 // uniqPointer
+	for i := 1; i < len(x); i++ {
+		if x[up] != x[i] {
+			up++
+			x[up] = x[i]
+		}
+	}
+	return x[:up+1]
+}