… example/client/go.sum
… go.sum
… hack/go.mod.default
… internal/slices/slices_test.go
… versions/HELM_VERSION
… versions/JAEGER_OPERATOR_VERSION
… versions/KUBELINTER_VERSION
… versions/PROMETHEUS_STACK_VERSION
… versions/PROTOBUF_VERSION
… versions/TELEPRESENCE_VERSION
… versions/YQ_VERSION