<!--- Provide a general summary of your changes in the Title above -->

### Description:
Since errors.Is was used to evaluate the equality of errors in errors.Join for the 0th and subsequent errors, performance problems occurred exponentially depending on the length of the errors.

By eliminating the evaluation of errors.Is, I have confirmed that performance can be improved.

```
❯ go test -count=3 -timeout=1h -run=NONE -bench . -benchmem
goos: linux
goarch: amd64
pkg: github.com/vdaas/vald/internal/errors
cpu: AMD Ryzen Threadripper 3990X 64-Core Processor
BenchmarkWrapLongData-128                     30          34398736 ns/op        138625268 B/op     88049 allocs/op
BenchmarkWrapLongData-128                     57          19898078 ns/op        137000936 B/op     47636 allocs/op
BenchmarkWrapLongData-128                     20          50354703 ns/op        140355950 B/op    130721 allocs/op
BenchmarkWrapShortData-128                 61291             16516 ns/op           10668 B/op         53 allocs/op
BenchmarkWrapShortData-128                 54051             19190 ns/op           10345 B/op         59 allocs/op
BenchmarkWrapShortData-128                142771              7767 ns/op           13588 B/op         30 allocs/op
BenchmarkJoinLongData-128                  36448             27727 ns/op           40963 B/op         71 allocs/op
BenchmarkJoinLongData-128                  54247             19699 ns/op           42938 B/op         48 allocs/op
BenchmarkJoinLongData-128                  34216             31447 ns/op           46316 B/op         75 allocs/op
BenchmarkJoinShortData-128               3351846               335.7 ns/op           312 B/op          0 allocs/op
BenchmarkJoinShortData-128               4782178               220.1 ns/op           309 B/op          0 allocs/op
BenchmarkJoinShortData-128               2409987               425.7 ns/op           323 B/op          1 allocs/op
BenchmarkStdWrapLongData-128                  38          27880411 ns/op        137966772 B/op     70084 allocs/op
BenchmarkStdWrapLongData-128                  38          26452982 ns/op        137953649 B/op     70085 allocs/op
BenchmarkStdWrapLongData-128                  50          20250260 ns/op        137281848 B/op     53914 allocs/op
BenchmarkStdWrapShortData-128              43831             23024 ns/op           10736 B/op         70 allocs/op
BenchmarkStdWrapShortData-128              71410             14347 ns/op           13066 B/op         47 allocs/op
BenchmarkStdWrapShortData-128              55992             18342 ns/op           10656 B/op         57 allocs/op
BenchmarkStdJoinLongData-128               93657             11985 ns/op           29765 B/op       1051 allocs/op
BenchmarkStdJoinLongData-128               13926             73825 ns/op           36024 B/op       1207 allocs/op
BenchmarkStdJoinLongData-128               11421             87668 ns/op           37639 B/op       1248 allocs/op
BenchmarkStdJoinShortData-128            1830832               549.7 ns/op           279 B/op          9 allocs/op
BenchmarkStdJoinShortData-128            1901421               552.7 ns/op           277 B/op          9 allocs/op
BenchmarkStdJoinShortData-128            1101970               917.7 ns/op           316 B/op         10 allocs/op
PASS
ok      github.com/vdaas/vald/internal/errors   873.150s

```
### Related Issue:

<!-- This project mainly accepts pull requests related to open issues -->
<!-- NOTE: If suggesting a new feature or change, please discuss it in an issue first -->
<!-- NOTE: If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!-- Please link to the issue here: -->

### Versions:

<!--- Please change the versions below along with your environment -->

- Go Version: 1.20.3
- Docker Version: 20.10.8
- Kubernetes Version: 1.22.0
- NGT Version: 2.0.9

### Checklist:

<!-- For completed items, change [ ] to [x]. -->
<!-- NOTE: these things are not required to open a PR and can be done afterwards / while the PR is open. -->

- [ ] I have read the [CONTRIBUTING](https://github.com/vdaas/vald/blob/main/CONTRIBUTING.md) document and completed [our CLA agreement](https://cla-assistant.io/vdaas/vald).
- [ ] I have checked open [Pull Requests](https://github.com/vdaas/vald/pulls) for the similar feature or fixes?

### Special notes for your reviewer:

<!-- Please tell us anything you would like to share to reviewers related this PR -->